* updated Allure version to the latest one;  * changed the way of getting browser parameter (instead of env var, reading testng xml parameters);  * moved screenshot taking staff into BaseListener;
