+++ framework/src/main/java/ue/ipscstore/core/BaseTest.java
@@ -1,10 +1,7 @@
 package ue.ipscstore.core;
 
-import org.openqa.selenium.OutputType;
-import org.openqa.selenium.TakesScreenshot;
-import org.openqa.selenium.remote.Augmenter;
-import org.testng.annotations.AfterMethod;
-import ru.yandex.qatools.allure.annotations.Attachment;
+import org.testng.ITestContext;
+import org.testng.annotations.Parameters;
 import ue.ipscstore.utils.PropertiesUtils;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
@@ -14,9 +11,7 @@ import org.testng.annotations.BeforeClass;
 import ue.ipscstore.utils.Browser;
 import ue.ipscstore.utils.WebDriverUtils;
 
-import java.io.File;
 import java.util.concurrent.TimeUnit;
-import java.util.logging.Logger;
 
 import static ue.ipscstore.utils.PropertiesUtils.getLongValue;
 
@@ -29,13 +24,13 @@ public class BaseTest {
         return WEB_DRIVER_CONTAINER.get();
     }
 
-    private static final Logger LOGGER = Logger.getLogger(BaseTest.class.getName());
-
     @BeforeClass
-    public void setUp() {
+    //@Parameters({"browser"})
+    public void setUp(ITestContext context) {
+        String browser = context.getCurrentXmlTest().getAllParameters().get("browser");
         WebDriver driver;
 
-        switch (Browser.getBrowser(System.getProperty("browser", "firefox"))) {
+        switch (Browser.getBrowser(browser/*System.getProperty("browser", "firefox")*/)) {
             case CHROME:
                 WebDriverUtils.setChromeDriverPath();
                 driver = new ChromeDriver();
@@ -51,17 +46,6 @@ public class BaseTest {
         WEB_DRIVER_CONTAINER.set(driver);
     }
 
-    @AfterMethod(alwaysRun = true)
-    public void takeScreenshot() {
-        LOGGER.info("Taking screenshot");
-        try {
-            WebDriver augmentedDriver = new Augmenter().augment(getDriver());
-            saveScreenshot("Screenshot", ((TakesScreenshot) augmentedDriver).getScreenshotAs(OutputType.BYTES));
-        } catch (Exception ex) {
-            LOGGER.severe("Unable to take screenshot: " + ex);
-        }
-    }
-
     @AfterClass
     public void tearDown(){
         if (getDriver() != null){
@@ -69,10 +53,4 @@ public class BaseTest {
             WEB_DRIVER_CONTAINER.remove();
         }
     }
-
-    @Attachment(value = "{0}", type = "image/png")
-    public byte[] saveScreenshot(String name, byte[] screenShot) {
-        return screenShot;
-    }
-
 }
