+++ pom.xml
@@ -16,7 +16,7 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <java.version>1.8</java.version>
-        <allure.version>1.4.16</allure.version>
+        <allure.version>1.4.23</allure.version>
         <aspectj.version>1.8.5</aspectj.version>
     </properties>
